new afk implementation
uses the idea that getPointerRotation never changes for a VR person except through the "lookAt" function.
If a player has 6 unique decimal pointer rotations, then they are not a VR player. Then use the old pointer position method of determining afk.
Otherwise treat them as a VR headset person and never kick them.

I also removed reliance on the "seated" player value, which gets corrupted when quickly changing seat colors.  Checking whether steam_id ~= nil is more reliable.
It will not afk grey blinded players.
Players that are grey, black, red, blue or blindfolded get free pointer jiggles while they are in those colours

since lookat can take a long time to stop changing the camera rotation(up to 5 seconds), wait until the pointer rotation has not changed for a whole second before measuring it.
Pointer rotations that are a whole number of degrees are not counted since VR players have these and they change with lookAt commands.  It is uncommon for a lookAt to have a decimal degree rotation and it is unlikely for a real person to land on an integer value when shifting the camera angle.