Fix "game is loading" message and broken state after loading an ended game
Added a new game state "ended". Settings are loaded when status is != -2 so that when a game is saved or rewound after it has ended, it loads correctly.
When the loading of a game errors, it permanently breaks,
so I made the "loading" state last for only 8 seconds after which a game can be restarted again.