Save player seats when rewinding within 60 seconds
Added the seating arrangement to the save state returned by onSave
Changed onLoad to implement the saved seating plan if less than 60 seconds have passed and a game was running.
Added a function that takes in any seating arrangement and implements it in the least number of seat changes.
Reworked how the code shuffles players to more easily make a seating plan table