Fix disappearing cluebox
Where there are many custom assets, the global xml never loads.
This happens for the host sometimes but it usually happens to the clients when they join the room.
As far as I can tell, the global xml on the client's computer is blank in memory and so setting properties of the empty xml doesn't do anything no matter what is tried.
The only solution I've found is to reload the whole xml from scratch.
This is mildly disruptive to the current codemasters, so I have it only happen at the following times:
Reset the global xml 5 seconds after the room loads to ensure it loaded.
Reset the global xml whenever a newly joining person sits in red or blue, because it probably doesn't exist for them.
Use the continually saved clue that was typed by the blue and red codemasters to repopulate the clue box upon reloading.