Improve saving for selected decks and let settings save when gameState.status == -2
Make the settings load from a previous save even when gameState.status == -2
Load the deck selections and the current deck page
Since downloading all decks takes a long time, wait 5 seconds to do so but set "fetchingDecks" to true
During that 5 seconds, the deck list is populated with dummy data that makes it look the same.
Delayed the search term from updating while the decks are being fetched
Made the page buttons responsive for the user even while fetching decks. There is no harm in this as page turning is done offline and the deck view will update when done fetching.