Completely offline deck handler
Custom assets take a reasonably long time to load for other users, on the order of seconds, so to display images properly we need to have all the images we are going to use already in the custom assets array.  And to do that we need to request all the decks ahead of time. Since we need to do this anyway, I decided to manage the decks offline so it is snappier.
Made pageNum 1-indexed instead of 0-indexed.
Request all the decks when loading room to use offline.
Create multiple web requests to download the decks
Make deck buttons and search use offline list.
Make reload button show message after decks are refreshed.
Make the deck reload button remove current decks visually.
If deck reloading fails, fall back on the 1-page saved deck list.