Fix some settings not saving/loading
Fix some settings not saving/loading

The "Enable Timer" button had the same id as the countdown clock, so I made it unique.
afk setting was not saving properly.
Moved the multivating table state adjustments to global so that they get called when the settings are applied from a saved room.
Paired up all the get and set settings functions